requirements.txt
streamlit
langchain
langchain-google
langchain-communinty
os
dotenv